Исправлена ошибка работы фильтра
